Re-establish standard project folder structure
https://discord.com/channels/1418641756383088680/1419052648560066660/1431775326962188358

This commit re-establishes a clear, standard folder structure for all teams to follow. This ensures all disciplines have a clear and predictable location for their work and keeps the project organized as it scales.

The structure separates source files from in-game prefabs:

`Assets/GameAssets/Art/` and `Assets/GameAssets/Audio/` are for source assets (e.g., models, textures, sound files).
`Assets/Prefabs/` is for the final, assembled in-game objects that are used in our scenes (e.g., `Characters`, `VFX`, `Environment` and 'UI').